+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - `0Kxy`: Retrigger ([#249]; thanks [@Speedvicio], [@Ravancloak])
   - `ESxx`: FM envelope reset ([#301]; thanks [@Zexxerd])
 - [#354] - VGM export for YM2610B ([#335]; thanks [@freq-mod], [@ToughkidDev])
+- [#345] - German notation system (thanks [@OPNA2608], [@wildmatsu], [@freq-mod])
 - [#332] - Qt6 support ([#298]; thanks [@OPNA2608])
 
 ### Changed
@@ -35,6 +36,7 @@
 [@Speedvicio]: https://github.com/Speedvicio
 [@RaijinXBlade]: https://github.com/RaijinXBlade
 [@ToughkidDev]: https://github.com/ToughkidDev
+[@wildmatsu]: https://github.com/wildmatsu
 
 [#328]: https://github.com/rerrahkr/BambooTracker/issues/328
 [#255]: https://github.com/rerrahkr/BambooTracker/issues/255
@@ -62,6 +64,7 @@
 [#366]: https://github.com/rerrahkr/BambooTracker/issues/366
 [#368]: https://github.com/rerrahkr/BambooTracker/pull/368
 [#367]: https://github.com/rerrahkr/BambooTracker/issues/367
+[#345]: https://github.com/rerrahkr/BambooTracker/issues/345
 
 ## v0.4.6 (2021-02-11)
 ### Added